Move generic.py into RimSort.utils
