+++ RimSort/views/game_configuration_panel.py
@@ -5,7 +5,7 @@ from PySide6.QtCore import QObject, Signal
 from RimSort.controllers.settings_controller import SettingsController
 from RimSort.models.dialogue import *
 from RimSort.utils.app_info import AppInfo
-from util.generic import *
+from RimSort.utils.generic import *
 from RimSort.windows.settings_panel import SettingsPanel
